refactor #4: 경로 적용 및 주석 제거
- common 컴포넌트 경로 수정
