+++ fe/public/index.html
@@ -1,5 +1,4 @@
 <!DOCTYPE html>
-<!-- NOTE(Jayden): CRA 기본 html에서 프로젝트에 필요없는 내용 삭제 및 title 수정 -->
 <html lang="ko">
   <head>
     <meta charset="utf-8" />
