+++ fe/src/common/Button/Button.stories.ts
@@ -1,5 +1,6 @@
 import type { Meta, StoryObj } from '@storybook/react';
-import Button from './Button';
+import Button from '../Button';
+import React from 'react';
 
 const meta = {
   title: 'Common/Button',
@@ -11,8 +12,13 @@ type Story = StoryObj<typeof meta>;
 
 export const Primary: Story = {
   args: {
+    isFlexible: false,
+    type: 'Contained',
     hasIcon: true,
+    condition: 'Enabled',
     size: 'Large',
+    color: 'Blue',
+    title: 'Button',
   },
   // NOTE(Jayden): onClick같은 경우는 argsType을 사용해야 test시에도 사용할 수 있다.
   argTypes: {
