+++ fe/src/common/Label/Label.stories.ts
@@ -1,5 +1,5 @@
 import type { Meta, StoryObj } from '@storybook/react';
-import Label from './Label';
+import Label from '../Label';
 
 const meta = {
   title: 'Main/Label',
