--- fe/src/common/Label/Label.tsx
@@ -1,21 +1,0 @@
-import React from 'react';
-
-interface Props {
-  labelName: string;
-  backgroundColor: string;
-  fontColor: string;
-}
-
-const Label: React.FC<Props> = ({ labelName, backgroundColor, fontColor }) => {
-  return (
-    // NOTE(Jayden): tailwindcss 이슈 - backgroundColor, fontColor는 style 객체로 넘겨줘야 함
-    <div
-      className={'mr-1 h-8 w-fit rounded-2xl px-6 py-2 text-xs font-semibold'}
-      style={{ backgroundColor, color: fontColor }}
-    >
-      {labelName}
-    </div>
-  );
-};
-
-export default Label;
