+++ fe/src/common/Logo/Logo.stories.ts
@@ -1,5 +1,5 @@
 import type { Meta, StoryObj } from '@storybook/react';
-import Logo from './Logo';
+import Logo from '../Logo';
 
 const meta = {
   title: 'Main/Logo',
