+++ fe/tsconfig.json
@@ -2,11 +2,7 @@
   "compilerOptions": {
     "baseUrl": "./",
     "target": "es5",
-    "lib": [
-      "dom",
-      "dom.iterable",
-      "esnext"
-    ],
+    "lib": ["dom", "dom.iterable", "esnext"],
     "allowJs": true,
     "skipLibCheck": true,
     "esModuleInterop": true,
@@ -22,7 +18,8 @@
     "jsx": "react-jsx",
     "paths": {
       "@common/*": ["src/common/*"],
-      "@components/*": ["src/components/*"]
+      "@components/*": ["src/components/*"],
+      "@assets/*": ["src/assets/*"],
     }
   },
   "include": ["src"]
